Move image_transforms artifacts
