+++ tests/artifacts/image_transforms/default_transforms.safetensors
