+++ tests/artifacts/image_transforms/single_transforms.safetensors
